--- tests/test_processor.py
@@ -1,83 +1,0 @@
-import kerastuner
-import numpy as np
-import tensorflow as tf
-
-from autokeras.hypermodel import processor
-
-
-def test_normalize():
-    normalize = processor.Normalize()
-    x_train = np.random.rand(100, 32, 32, 3)
-    dataset = tf.data.Dataset.from_tensor_slices(x_train)
-    normalize.set_hp(kerastuner.HyperParameters())
-    for x in dataset:
-        normalize.update(x)
-
-    def map_func(x):
-        return tf.py_function(normalize.transform,
-                              inp=[x],
-                              Tout=(tf.float64,))
-
-    new_dataset = dataset.map(map_func)
-    assert isinstance(new_dataset, tf.data.Dataset)
-
-
-def test_sequence():
-    texts = ['The cat sat on the mat.',
-             'The dog sat on the log.',
-             'Dogs and cats living together.']
-    tokenize = processor.TextToIntSequence()
-    dataset = tf.data.Dataset.from_tensor_slices(texts)
-    tokenize.set_hp(kerastuner.HyperParameters())
-    for x in dataset:
-        tokenize.update(x)
-
-    def map_func(x):
-        return tf.py_function(tokenize.transform,
-                              inp=[x],
-                              Tout=(tf.int64,))
-
-    new_dataset = dataset.map(map_func)
-    for _ in new_dataset:
-        break
-    assert isinstance(new_dataset, tf.data.Dataset)
-
-
-def test_ngram():
-    texts = ['The cat sat on the mat.',
-             'The dog sat on the log.',
-             'Dogs and cats living together.']
-    tokenize = processor.TextToNgramVector()
-    dataset = tf.data.Dataset.from_tensor_slices(texts)
-    tokenize.set_hp(kerastuner.HyperParameters())
-    for x in dataset:
-        tokenize.update(x)
-
-    def map_func(x):
-        return tf.py_function(tokenize.transform,
-                              inp=[x],
-                              Tout=(tf.float64,))
-
-    new_dataset = dataset.map(map_func)
-    assert isinstance(new_dataset, tf.data.Dataset)
-
-    
-def test_ngram():
-    texts = ['The cat sat on the mat.',
-             'The dog sat on the log.',
-             'Dogs and cats living together.']
-    tokenize = processor.TextToNgramVector()
-    dataset = tf.data.Dataset.from_tensor_slices(texts)
-    tokenize.set_hp(kerastuner.HyperParameters())
-    for x in dataset:
-        tokenize.update(x)
-    tokenize.finalize()
-    def map_func(x):
-        return tf.py_function(tokenize.transform,
-                              inp=[x],
-                              Tout=(tf.float64,))
-
-    new_dataset = dataset.map(map_func)
-    for _ in new_dataset:
-        break
-    assert isinstance(new_dataset, tf.data.Dataset)
